5.1.6 Перенесет функции для рендеринга из utils.js в utils/render.js
