+++ src/utils/render.js
@@ -1,0 +1,24 @@
+const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  AFTEREND: `afterend`,
+  BEFOREEND: `beforeend`,
+};
+
+const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
+
+const render = (container, element, place) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN: container.prepend(element);
+      break;
+    case RenderPosition.BEFOREEND: container.append(element);
+      break;
+    case RenderPosition.AFTEREND: container.after(element);
+  }
+};
+
+export {RenderPosition, createElement, render};
